// frontend/src/types/profile.ts
export type Gender = "Male" | "Female" | "Other"

export type SexualOrientation =
    | "Straight" | "Gay" | "Lesbian" | "Bisexual" | "Asexual"
    | "Demisexual" | "Pansexual" | "Queer" | "Questioning"

export type RelationshipGoal =
    | "LongTermPartner" | "LongTermOpenToShort" | "ShortTermOpenToLong"
    | "ShortTermFun" | "NewFriends" | "StillFiguringItOut"

export type FamilyPlan =
    | "WantChildren" | "DontWantChildren" | "HaveChildrenWantMore"
    | "HaveChildrenDontWantMore" | "NotSureYet"

export type CommunicationStyle =
    | "BigTimeTexter" | "PhoneCaller" | "VideoChatter" | "BadTexter" | "BetterInPerson"

export type LoveStyle =
    | "ThoughtfulGestures" | "Presents" | "Touch" | "Compliments" | "TimeTogether"

export type PetPreference =
    | "Dog" | "Cat" | "Reptile" | "Amphibian" | "Bird" | "Fish"
    | "DontHaveButLove" | "Allergic" | "NoPets" | "WantAPet" | "Other"

export type DrinkingHabit =
    | "NotForMe" | "SoberCurious" | "OnSpecialOccasions" | "SociallyOnWeekends" | "MostNights"

export type SmokingHabit =
    | "NonSmoker" | "SocialSmoker" | "SmokerWhenDrinking" | "Smoker" | "TryingToQuit"

export type WorkoutHabit = "Everyday" | "Often" | "Sometimes" | "Never"